pyproject.toml
[project]
name = "mcpd-plugins"
dynamic = ["version"]
description = "mcpd plugins Python SDK"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.76.0",
    "protobuf>=6.33.0,<7",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.76.0",
    "debugpy>=1.8.17",
]
lint = [
    "pre-commit>=4.3.0",
    "ruff>=0.14.2",
]
tests = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "setuptools>=80.9.0",
    "setuptools-scm>=9.2.2",
]
build = [
    "pyinstaller>=6.16.0",
    "nuitka>=2.8.4",
]
all = [
    { include-group = "dev" },
    { include-group = "lint" },
    { include-group = "tests" },
]

[build-system]
requires = ["setuptools>=80.9.0", "setuptools_scm[toml]>=9.2.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests", "tests.*"]
namespaces = false

[tool.setuptools.package-data]
mcpd_plugins = ["py.typed"]

[tool.setuptools_scm]
write_to = "src/mcpd_plugins/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
log_cli_level = "INFO"
log_level = "INFO"
